Implement Planner support for Union ALL
The Union ALL planner support uses a `Union` operator in the `LogicalPlanner`
which generates a `UnionPlan`. The plan consists of a sub plan for each side of
the union. When the union is planned, both sides are executed as a branch of the
execution tree. The result is fed to the `MergePhase` which receives data from
both sub plans and merges them into buckets encoded by the node id and the input
id set during planning.

In `MergePhase` the number of different inputs was added to ensure that a merge
with multiple inputs can be detected. The input id is encoded at the beginning
of the 32bit bucket id.