Adjust PageDownstreamContext for new bucket id generation
Previously we assumed that bucket indices would be sequentially generated from 0
to numBuckets - 1. This is not the case anymore because we encode the input
phase id in the bucket id. Other operations like NestedLoop should not be
concerned with the new bucket indices as they're execution behavior remains the
same. The changes make the actual index values transparent to the
`PageDownstreamContext`.